requirements.txt
arrow==0.10.0
falcon==1.0.0
hug==2.2.0
iniherit==0.3.6
python-dateutil==2.6.0
python-mimeparse==1.6.0
python-slugify==1.2.1
requests==2.12.4
six==1.10.0
Unidecode==0.4.20
yoyo-migrations==5.0.5
